Conflicts:
	authentication/authentication-service/src/main/java/io/mosip/authentication/service/impl/otpgen/service/OTPServiceImpl.java

Former-commit-id: 64e21acf5e9f4d726a0e1fe2828a96adb2b51ced
Former-commit-id: 829acd598312aa2fd6de3d6d6e3c7c7e5a560422

+++ registration/registration-client/src/main/java/io/mosip/registration/controller/reg/PacketUploadController.java
@@ -6,23 +6,29 @@ import static io.mosip.registration.constants.RegistrationConstants.APPLICATION_
 import java.io.File;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.WeakHashMap;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 
+import io.mosip.kernel.core.exception.ExceptionUtils;
 import io.mosip.kernel.core.logger.spi.Logger;
+import io.mosip.kernel.core.util.DateUtils;
 import io.mosip.kernel.core.util.exception.JsonProcessingException;
 import io.mosip.registration.config.AppConfig;
 import io.mosip.registration.constants.AuditEvent;
+import io.mosip.registration.constants.AuditReferenceIdTypes;
 import io.mosip.registration.constants.Components;
 import io.mosip.registration.constants.RegistrationClientStatusCode;
 import io.mosip.registration.constants.RegistrationConstants;
+import io.mosip.registration.constants.RegistrationUIConstants;
 import io.mosip.registration.context.SessionContext;
 import io.mosip.registration.controller.BaseController;
 import io.mosip.registration.dto.PacketStatusDTO;
+import io.mosip.registration.dto.RegistrationPacketSyncDTO;
+import io.mosip.registration.dto.ResponseDTO;
 import io.mosip.registration.dto.SyncRegistrationDTO;
 import io.mosip.registration.entity.Registration;
 import io.mosip.registration.exception.RegBaseCheckedException;
@@ -30,6 +36,7 @@ import io.mosip.registration.exception.RegBaseUncheckedException;
 import io.mosip.registration.exception.RegistrationExceptionConstants;
 import io.mosip.registration.service.packet.PacketUploadService;
 import io.mosip.registration.service.sync.PacketSynchService;
+import io.mosip.registration.util.healthcheck.RegistrationAppHealthCheckUtil;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.concurrent.Service;
@@ -46,6 +53,12 @@ import javafx.scene.control.cell.PropertyValueFactory;
 public class PacketUploadController extends BaseController {
 
 	@FXML
+	private ProgressIndicator progressIndicator;
+
+	@Autowired
+	private PacketUploadService packetUploadService;
+
+	@FXML
 	private TableColumn<PacketStatusDTO, String> fileNameColumn;
 
 	@FXML
@@ -53,12 +66,6 @@ public class PacketUploadController extends BaseController {
 
 	@FXML
 	private TableView<PacketStatusDTO> table;
-
-	@FXML
-	private ProgressIndicator progressIndicator;
-
-	@Autowired
-	private PacketUploadService packetUploadService;
 
 	@Autowired
 	private PacketSynchService packetSynchService;
@@ -73,35 +80,46 @@ public class PacketUploadController extends BaseController {
 	 * 
 	 */
 	public void syncAndUploadPacket() {
+
 		LOGGER.info("REGISTRATION - SYNCH_PACKETS_AND_PUSH_TO_SERVER - PACKET_UPLOAD_CONTROLLER", APPLICATION_NAME,
 				APPLICATION_ID, "Sync the packets and push it to the server");
 		table.getItems().clear();
 		table.refresh();
 		service.reset();
 		try {
-			String packetSyncStatus = packetSync();
-			progressIndicator.progressProperty().bind(service.progressProperty());
-			service.start();
-			service.setOnSucceeded(new EventHandler<WorkerStateEvent>() {
-				@Override
-				public void handle(WorkerStateEvent t) {
-					String status = service.getValue();
-					if (!status.equals(RegistrationConstants.EMPTY)) {
-						String[] displayStatus = status.split("-");
-						if (RegistrationConstants.PACKET_SYNC_ERROR.equals(displayStatus[0])) {
-							if (!RegistrationConstants.EMPTY.equals(packetSyncStatus)) {
-								generateAlert(displayStatus[0], displayStatus[1] + " " + packetSyncStatus);
+			if (RegistrationAppHealthCheckUtil.isNetworkAvailable()) {
+				String packetSyncStatus = packetSync();
+
+				auditFactory.audit(AuditEvent.UPLOAD_PACKET, Components.UPLOAD_PACKET,
+						SessionContext.userContext().getUserId(), AuditReferenceIdTypes.USER_ID.getReferenceTypeId());
+
+				progressIndicator.progressProperty().bind(service.progressProperty());
+				service.start();
+				service.setOnSucceeded(new EventHandler<WorkerStateEvent>() {
+					@Override
+					public void handle(WorkerStateEvent t) {
+						String status = service.getValue();
+						if (!status.equals(RegistrationConstants.EMPTY)) {
+							String[] displayStatus = status.split("-");
+							if (RegistrationConstants.PACKET_SYNC_ERROR.equals(displayStatus[0])) {
+								if (!RegistrationConstants.EMPTY.equals(packetSyncStatus)) {
+									generateAlert(displayStatus[0], displayStatus[1] + " " + packetSyncStatus);
+								} else {
+									generateAlert(displayStatus[0], displayStatus[1]);
+								}
 							} else {
 								generateAlert(displayStatus[0], displayStatus[1]);
 							}
-						} else {
-							generateAlert(displayStatus[0], displayStatus[1]);
-						}
 
+						}
 					}
-				}
-			});
+				});
+			} else {
+				generateAlert(RegistrationConstants.ERROR, RegistrationUIConstants.NETWORK_ERROR);
+			}
 		} catch (RegBaseCheckedException checkedException) {
+			LOGGER.info("REGISTRATION - UPLOAD_ERROR - PACKET_UPLOAD_CONTROLLER", APPLICATION_NAME, APPLICATION_ID,
+					checkedException.getMessage() + ExceptionUtils.getStackTrace(checkedException));
 			generateAlert(RegistrationConstants.ERROR, checkedException.getErrorText());
 		}
 
@@ -118,12 +136,13 @@ public class PacketUploadController extends BaseController {
 				APPLICATION_ID, "Sync the packets to the server");
 		String syncErrorStatus = "";
 		try {
-			auditFactory.audit(AuditEvent.SYNC_SERVER, Components.PACKET_SYNC, "Sync the packets status to the server",
-					SessionContext.userContext().getUserId(),
-					RegistrationConstants.PACKET_SYNC_REF_ID);
+			auditFactory.audit(AuditEvent.UPLOAD_PACKET, Components.UPLOAD_PACKET,
+					SessionContext.userContext().getUserId(), AuditReferenceIdTypes.USER_ID.getReferenceTypeId());
+
 			List<Registration> packetsToBeSynched = packetSynchService.fetchPacketsToBeSynched();
 			List<SyncRegistrationDTO> syncDtoList = new ArrayList<>();
-			Object response = null;
+			List<Registration> synchedPackets = new ArrayList<>();
+			ResponseDTO responseDTO = new ResponseDTO();
 			if (!packetsToBeSynched.isEmpty()) {
 				for (Registration packetToBeSynch : packetsToBeSynched) {
 					SyncRegistrationDTO syncDto = new SyncRegistrationDTO();
@@ -135,78 +154,50 @@ public class PacketUploadController extends BaseController {
 					syncDto.setSyncType(RegistrationConstants.PACKET_STATUS_SYNC_TYPE);
 					syncDtoList.add(syncDto);
 				}
-				response = packetSynchService.syncPacketsToServer(syncDtoList);
+				RegistrationPacketSyncDTO registrationPacketSyncDTO = new RegistrationPacketSyncDTO();
+				registrationPacketSyncDTO.setRequestTimestamp(DateUtils.getUTCCurrentDateTimeString());
+				registrationPacketSyncDTO.setSyncRegistrationDTOs(syncDtoList);
+				registrationPacketSyncDTO.setId(RegistrationConstants.PACKET_SYNC_STATUS_ID);
+				registrationPacketSyncDTO.setVersion(RegistrationConstants.PACKET_SYNC_VERSION);
+				responseDTO = packetSynchService.syncPacketsToServer(registrationPacketSyncDTO);
 			}
-			if (response != null) {
-				packetsToBeSynched.forEach(registration -> {
-					if (registration.getServerStatusCode() != null) {
-						if (registration.getServerStatusCode()
+			if (responseDTO != null && responseDTO.getSuccessResponseDTO() != null) {
+
+				for (Registration registration : packetsToBeSynched) {
+					String status = (String) responseDTO.getSuccessResponseDTO().getOtherAttributes()
+							.get(registration.getId());
+					if (status != null && status.equalsIgnoreCase(RegistrationConstants.SUCCESS)) {
+
+						registration.setClientStatusCode(RegistrationClientStatusCode.META_INFO_SYN_SERVER.getCode());
+
+						if (registration.getServerStatusCode() != null && registration.getServerStatusCode()
 								.equals(RegistrationClientStatusCode.RE_REGISTER.getCode())) {
+
 							String ackFileName = registration.getAckFilename();
 							int lastIndex = ackFileName.indexOf(RegistrationConstants.ACKNOWLEDGEMENT_FILE);
 							String packetPath = ackFileName.substring(0, lastIndex);
 							File packet = new File(packetPath + RegistrationConstants.ZIP_FILE_EXTENSION);
 							if (packet.exists() && packet.delete()) {
 								registration.setClientStatusCode(RegistrationClientStatusCode.DELETED.getCode());
-							} else {
-								registration.setClientStatusCode(
-										RegistrationClientStatusCode.META_INFO_SYN_SERVER.getCode());
 							}
 						}
-					} else {
-						registration.setClientStatusCode(RegistrationClientStatusCode.META_INFO_SYN_SERVER.getCode());
+						synchedPackets.add(registration);
 					}
-				});
-				packetSynchService.updateSyncStatus(packetsToBeSynched);
+				}
+				packetSynchService.updateSyncStatus(synchedPackets);
 			}
-		} catch (RegBaseUncheckedException | RegBaseCheckedException | JsonProcessingException | URISyntaxException e) {
+		} catch (RegBaseCheckedException | JsonProcessingException | URISyntaxException exception) {
 			LOGGER.error("REGISTRATION - SYNCH_PACKETS_TO_SERVER - PACKET_UPLOAD_CONTROLLER", APPLICATION_NAME,
-					APPLICATION_ID, "Error while Synching packets to the server");
-			if (e instanceof RegBaseUncheckedException) {
+					APPLICATION_ID,
+					"Error while Synching packets to the server" + ExceptionUtils.getStackTrace(exception));
 
-				throw new RegBaseCheckedException(
-						RegistrationExceptionConstants.REG_PACKET_SYNC_EXCEPTION.getErrorCode(),
-						RegistrationExceptionConstants.REG_PACKET_SYNC_EXCEPTION.getErrorMessage());
-			} else {
-				syncErrorStatus = e.getMessage();
-			}
+			syncErrorStatus = exception.getMessage();
+
+		} catch (RegBaseUncheckedException regBaseUncheckedException) {
+			throw new RegBaseCheckedException(RegistrationExceptionConstants.REG_PACKET_SYNC_EXCEPTION.getErrorCode(),
+					RegistrationExceptionConstants.REG_PACKET_SYNC_EXCEPTION.getErrorMessage());
 		}
 		return syncErrorStatus;
-	}
-
-	/**
-	 * To display the Uploaded packet details in UI
-	 * 
-	 * @param tableData
-	 */
-	private void displayData(List<PacketStatusDTO> tableData) {
-		LOGGER.info("REGISTRATION - DISPLAY_DATA - PACKET_UPLOAD_CONTROLLER", APPLICATION_NAME, APPLICATION_ID,
-				"To display all the ui data");
-		fileNameColumn.setCellValueFactory(new PropertyValueFactory<>("fileName"));
-		uploadStatusColumn.setCellValueFactory(new PropertyValueFactory<>("uploadStatus"));
-
-		ObservableList<PacketStatusDTO> list = FXCollections.observableArrayList(tableData);
-		table.setItems(list);
-	}
-
-	/**
-	 * To populate the data for the UI table
-	 * 
-	 * @param verifiedPackets
-	 * @return
-	 */
-	private List<PacketStatusDTO> populateTableData(Map<String, String> packetStatus) {
-		LOGGER.info("REGISTRATION - POPULATE_UI_TABLE_DATA - PACKET_UPLOAD_CONTROLLER", APPLICATION_NAME,
-				APPLICATION_ID, "Populating the table data with the Updated details");
-		List<PacketStatusDTO> listUploadStatus = new ArrayList<>();
-		packetStatus.forEach((id, status) -> {
-			PacketStatusDTO packetUploadStatusDTO = new PacketStatusDTO();
-			packetUploadStatusDTO.setUploadStatus(status);
-			packetUploadStatusDTO.setFileName(id);
-			listUploadStatus.add(packetUploadStatusDTO);
-
-		});
-		return listUploadStatus;
 	}
 
 	/**
@@ -234,13 +225,12 @@ public class PacketUploadController extends BaseController {
 							APPLICATION_NAME, APPLICATION_ID, "Handling all the packet upload activities");
 					List<Registration> synchedPackets = packetUploadService.getSynchedPackets();
 					List<Registration> packetUploadList = new ArrayList<>();
-					Map<String, String> tableMap = new HashMap<String, String>();
 					String status = "";
+					Map<String, String> tableMap = new WeakHashMap<>();
 					if (!synchedPackets.isEmpty()) {
 						auditFactory.audit(AuditEvent.PACKET_UPLOAD, Components.PACKET_UPLOAD,
-								"Upload packets to the server",
-								SessionContext.userContext().getUserId(),
-								RegistrationConstants.PACKET_UPLOAD_REF_ID);
+								SessionContext.userContext().getUserId(), RegistrationConstants.PACKET_UPLOAD_REF_ID);
+
 						progressIndicator.setVisible(true);
 						for (int i = 0; i < synchedPackets.size(); i++) {
 							Registration synchedPacket = synchedPackets.get(i);
@@ -251,7 +241,7 @@ public class PacketUploadController extends BaseController {
 							File packet = new File(packetPath + RegistrationConstants.ZIP_FILE_EXTENSION);
 							try {
 								if (packet.exists()) {
-									if ((("resend".equals(synchedPacket.getServerStatusCode())
+									if ((("RESEND".equals(synchedPacket.getServerStatusCode())
 											&& synchedPacket.getServerStatusTimestamp()
 													.compareTo(synchedPacket.getUploadTimestamp()) == 1)
 											|| RegistrationClientStatusCode.META_INFO_SYN_SERVER.getCode()
@@ -260,9 +250,9 @@ public class PacketUploadController extends BaseController {
 													.equals(synchedPacket.getClientStatusCode())
 											|| "E".equals(synchedPacket.getFileUploadStatus())) && packet.exists()) {
 
-										Object response = packetUploadService.pushPacket(packet);
-										String responseCode = response.toString();
-										if (responseCode.equals("PACKET_UPLOADED_TO_VIRUS_SCAN")) {
+										ResponseDTO response = packetUploadService.pushPacket(packet);
+										if (response.getSuccessResponseDTO() != null) {
+
 											synchedPacket.setClientStatusCode(
 													RegistrationClientStatusCode.UPLOADED_SUCCESSFULLY.getCode());
 											synchedPacket.setFileUploadStatus(
@@ -271,44 +261,53 @@ public class PacketUploadController extends BaseController {
 											tableMap.put(synchedPacket.getId(),
 													RegistrationConstants.PACKET_UPLOAD_SUCCESS);
 
-										} else {
-											synchedPacket.setFileUploadStatus(
-													RegistrationClientStatusCode.UPLOAD_ERROR_STATUS.getCode());
-											packetUploadList.add(synchedPacket);
-											tableMap.put(synchedPacket.getId(), "Error");
+										} else if (response.getErrorResponseDTOs() != null) {
+											String errMessage = response.getErrorResponseDTOs().get(0).getMessage();
+											if (errMessage.contains(RegistrationConstants.PACKET_DUPLICATE)) {
+
+												tableMap.put(synchedPacket.getId(), "Error(Duplicate Packet)");
+												synchedPacket.setClientStatusCode(
+														RegistrationClientStatusCode.UPLOADED_SUCCESSFULLY.getCode());
+												synchedPacket.setFileUploadStatus(
+														RegistrationClientStatusCode.UPLOAD_SUCCESS_STATUS.getCode());
+												packetUploadList.add(synchedPacket);
+
+											} else {
+												synchedPacket.setFileUploadStatus(
+														RegistrationClientStatusCode.UPLOAD_ERROR_STATUS.getCode());
+												packetUploadList.add(synchedPacket);
+												tableMap.put(synchedPacket.getId(), "Error");
+											}
 										}
+
 									}
 								} else {
 									tableMap.put(synchedPacket.getId(), "Error(Packet not available)");
 								}
 
-							} catch (URISyntaxException e) {
+							} catch (URISyntaxException uriSyntaxException) {
 
 								LOGGER.error("REGISTRATION - HANDLE_PACKET_UPLOAD_URI_ERROR - PACKET_UPLOAD_CONTROLLER",
-										APPLICATION_NAME, APPLICATION_ID, "Error in uri syntax");
+										APPLICATION_NAME, APPLICATION_ID,
+										"Error in uri syntax" + ExceptionUtils.getStackTrace(uriSyntaxException));
 								status = "Error-Unable to push packets to the server.";
-							} catch (RegBaseCheckedException e) {
+							} catch (RegBaseCheckedException regBaseCheckedException) {
 								LOGGER.error("REGISTRATION - HANDLE_PACKET_UPLOAD_ERROR - PACKET_UPLOAD_CONTROLLER",
-										APPLICATION_NAME, APPLICATION_ID, "Error while pushing packets to the server");
-								if(e.getMessage().contains(RegistrationConstants.PACKET_DUPLICATE)) {
-									tableMap.put(synchedPacket.getId(), "Error(Duplicate Packets)");
-									synchedPacket.setClientStatusCode(RegistrationClientStatusCode.UPLOADED_SUCCESSFULLY.getCode());
-									synchedPacket.setFileUploadStatus(
-											RegistrationClientStatusCode.UPLOAD_SUCCESS_STATUS.getCode());
-								} else {
-									synchedPacket.setFileUploadStatus(
-											RegistrationClientStatusCode.UPLOAD_ERROR_STATUS.getCode());
-									tableMap.put(synchedPacket.getId(), "Error(" + e.getErrorTexts().toString() + ")");
-								}
+										APPLICATION_NAME, APPLICATION_ID, "Error while pushing packets to the server"
+												+ ExceptionUtils.getStackTrace(regBaseCheckedException));
+
+								synchedPacket.setFileUploadStatus(
+										RegistrationClientStatusCode.UPLOAD_ERROR_STATUS.getCode());
+								tableMap.put(synchedPacket.getId(), "Error(Service Error)");
 								packetUploadList.add(synchedPacket);
 								synchedPacket.setUploadCount((short) (synchedPacket.getUploadCount() + 1));
-								
 
-							} catch (RuntimeException e) {
+							} catch (RuntimeException runtimeException) {
 								LOGGER.error(
 										"REGISTRATION - HANDLE_PACKET_UPLOAD_RUNTIME_ERROR - PACKET_UPLOAD_CONTROLLER",
 										APPLICATION_NAME, APPLICATION_ID,
-										"Run time error while connecting to the server");
+										"Run time error while connecting to the server"
+												+ ExceptionUtils.getStackTrace(runtimeException));
 								if (i == 0) {
 									status = "Error-Unable to push packets to the server.";
 								} else if (i > 0) {
@@ -349,10 +348,46 @@ public class PacketUploadController extends BaseController {
 				"Exporting the Synched the packets");
 
 		List<Registration> exportedPackets = packetExportController.packetExport();
-		Map<String, String> exportedPacketMap = new HashMap<>();
+		Map<String, String> exportedPacketMap = new WeakHashMap<>();
 		exportedPackets.forEach(regPacket -> {
 			exportedPacketMap.put(regPacket.getId(), RegistrationClientStatusCode.EXPORT.getCode());
 		});
 		displayData(populateTableData(exportedPacketMap));
+	}
+
+	/**
+	 * To display the Uploaded packet details in UI
+	 * 
+	 * @param tableData
+	 */
+	private void displayData(List<PacketStatusDTO> tableData) {
+		
+		LOGGER.info("REGISTRATION - DISPLAY_DATA - PACKET_UPLOAD_CONTROLLER", APPLICATION_NAME, APPLICATION_ID,
+				"To display all the ui data");
+		fileNameColumn.setCellValueFactory(new PropertyValueFactory<>("fileName"));
+		uploadStatusColumn.setCellValueFactory(new PropertyValueFactory<>("uploadStatus"));
+
+		ObservableList<PacketStatusDTO> list = FXCollections.observableArrayList(tableData);
+		table.setItems(list);
+	}
+
+	/**
+	 * To populate the data for the UI table
+	 * 
+	 * @param verifiedPackets
+	 * @return
+	 */
+	private List<PacketStatusDTO> populateTableData(Map<String, String> packetStatus) {
+		LOGGER.info("REGISTRATION - POPULATE_UI_TABLE_DATA - PACKET_UPLOAD_CONTROLLER", APPLICATION_NAME,
+				APPLICATION_ID, "Populating the table data with the Updated details");
+		List<PacketStatusDTO> listUploadStatus = new ArrayList<>();
+		packetStatus.forEach((id, status) -> {
+			PacketStatusDTO packetUploadStatusDTO = new PacketStatusDTO();
+			packetUploadStatusDTO.setUploadStatus(status);
+			packetUploadStatusDTO.setFileName(id);
+			listUploadStatus.add(packetUploadStatusDTO);
+
+		});
+		return listUploadStatus;
 	}
 }
